feat(backend): move the cost reporting surface into the optional package
The routes and their response schemas now live in the separately installed
package and mount through the existing admin-extension hook, so an unlicensed
deployment does not serve them. The middleware that records usage per run, the
repository it writes through, and the models are untouched in core.

The panel exposes a router and nothing else. It returns no nav item on purpose:
the host frontend already carries an edition-gated Insights link, and manifest
nav items render as links into an iframe surface meant for panels that ship
their own UI — one here would have produced a duplicate link loading this JSON
API in an iframe. The router keeps the parent's `admin` tag, which direct
mounting would otherwise drop and silently regroup the generated schema.

Tests split three ways rather than gated wholesale. The old file mixed four
HTTP cases with four that drive BillingRepository directly; a module-level
importorskip would have dropped the repository coverage that keeping the
repository in core exists to preserve. Seeding helpers moved to
tests/e2e/billing_fixtures.py so both halves share them.

Verified all three environment states behave as designed:
- default install — repository tests 4 passed 0 skipped, gated module skips,
  and a new module asserts both the old and new paths 404 for an authenticated
  admin and that the schema advertises neither
- installed with a key — gated HTTP tests 4 passed at the new mount point,
  which is also what confirms the derived prefix
- installed without a key — collection aborts, as it already did

Also retargets a consumer the plan review found: a real-LLM test polled the
endpoint, and the planned sweep excludes that marker, so the nightly lane would
have started 404ing unnoticed. Its invariant straddles the boundary, so it now
asserts the usage row was written — the part that is core — and leaves reading
it back to the gated module, which needs no live model call. 2 passed.
