+++ backend/ee/src/cubeplex_ee/cost/__init__.py
@@ -1,0 +1,5 @@
+"""Cost reporting: the read surface over the usage rows the core writes."""
+
+from cubeplex_ee.cost.panel import CostPanel
+
+__all__ = ["CostPanel"]
