+++ backend/tests/e2e/test_billing_cost_api.py
@@ -1,188 +1,14 @@
-"""E2E tests for the redesigned /admin/cost/* surface and timeseries repo."""
+"""E2E: BillingRepository timeseries aggregation. Core surface, default install."""
 
-from collections.abc import AsyncIterator
-from contextlib import asynccontextmanager
-from datetime import UTC, datetime, timedelta
+from datetime import UTC, datetime
 from typing import Any
 
-import httpx
 import pytest
-from sqlalchemy import delete, select
-from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
-from sqlalchemy.pool import NullPool
 
-from cubeplex.db.engine import _build_database_url
-from cubeplex.models.billing import BillingEvent, LlmBillingEvent
-from cubeplex.models.conversation import Conversation
-from cubeplex.models.organization import Organization
-from cubeplex.models.user import User
-from cubeplex.models.workspace import Workspace
 from cubeplex.repositories import BillingRepository
-from tests.e2e.conftest import DEFAULT_ORG_ID, DEFAULT_WS_ID
+from tests.e2e.billing_fixtures import _db_session, _seed_events
 
 pytestmark = pytest.mark.e2e
-
-
-@asynccontextmanager
-async def _db_session() -> AsyncIterator[AsyncSession]:
-    """Create a direct AsyncSession to the test DB (NullPool, no connection sharing)."""
-    engine = create_async_engine(_build_database_url(), poolclass=NullPool)
-    maker = async_sessionmaker(engine, class_=AsyncSession, expire_on_commit=False)
-    try:
-        async with maker() as session:
-            yield session
-    finally:
-        await engine.dispose()
-
-
-async def _ensure_org(session: AsyncSession, org_id: str) -> None:
-    existing = (
-        await session.execute(select(Organization).where(Organization.id == org_id))
-    ).scalar_one_or_none()
-    if existing is None:
-        session.add(
-            Organization(
-                id=org_id,
-                name=f"Test {org_id}",
-                slug=org_id.replace("_", "-").lower()[:32],
-            )
-        )
-        await session.commit()
-
-
-async def _ensure_workspace(session: AsyncSession, *, ws_id: str, org_id: str) -> None:
-    existing = (
-        await session.execute(select(Workspace).where(Workspace.id == ws_id))
-    ).scalar_one_or_none()
-    if existing is None:
-        session.add(Workspace(id=ws_id, org_id=org_id, name=f"Test {ws_id}"))
-        await session.commit()
-
-
-async def _ensure_user(
-    session: AsyncSession,
-    *,
-    user_id: str,
-    display_name: str | None = None,
-) -> None:
-    existing = (await session.execute(select(User).where(User.id == user_id))).scalar_one_or_none()
-    if existing is None:
-        session.add(
-            User(
-                id=user_id,
-                email=f"{user_id}@test-billing-cost.local",
-                hashed_password="x",
-                is_active=True,
-                display_name=display_name,
-            )
-        )
-        await session.commit()
-    elif display_name is not None and existing.display_name != display_name:
-        existing.display_name = display_name
-        session.add(existing)
-        await session.commit()
-
-
-async def _ensure_conversation(
-    session: AsyncSession,
-    *,
-    conv_id: str,
-    org_id: str,
-    ws_id: str,
-    user_id: str,
-) -> None:
-    existing = (
-        await session.execute(select(Conversation).where(Conversation.id == conv_id))
-    ).scalar_one_or_none()
-    if existing is None:
-        session.add(
-            Conversation(
-                id=conv_id,
-                org_id=org_id,
-                workspace_id=ws_id,
-                creator_user_id=user_id,
-                title="Test conv",
-            )
-        )
-        await session.commit()
-
-
-async def _seed_events(
-    session: AsyncSession,
-    *,
-    org_id: str,
-    rows: list[dict[str, Any]],
-) -> None:
-    """Insert billing rows after ensuring FK parents (org/ws/user/conversation) exist.
-
-    Deletes any prior billing rows for ``org_id`` first so reruns are idempotent
-    against a shared dev/test database.
-    """
-    # Clean prior billing rows for this test org (child first to satisfy FK).
-    prior_ids = (
-        (await session.execute(select(BillingEvent.id).where(BillingEvent.org_id == org_id)))
-        .scalars()
-        .all()
-    )
-    if prior_ids:
-        await session.execute(
-            delete(LlmBillingEvent).where(LlmBillingEvent.billing_event_id.in_(prior_ids))
-        )
-        await session.execute(delete(BillingEvent).where(BillingEvent.org_id == org_id))
-        await session.commit()
-    await _ensure_org(session, org_id)
-    # Track which parents we've ensured to keep seeding fast.
-    seen_ws: set[str] = set()
-    seen_users: set[str] = set()
-    seen_conv: set[str] = set()
-    for r in rows:
-        ws_id = str(r["workspace_id"])
-        user_id = str(r["user_id"])
-        conv_id = str(r.get("conversation_id", f"conv-seed-{org_id[:8]}"))
-        if ws_id not in seen_ws:
-            await _ensure_workspace(session, ws_id=ws_id, org_id=org_id)
-            seen_ws.add(ws_id)
-        if user_id not in seen_users:
-            await _ensure_user(
-                session,
-                user_id=user_id,
-                display_name=r.get("user_display_name"),  # type: ignore[arg-type]
-            )
-            seen_users.add(user_id)
-        if conv_id not in seen_conv:
-            await _ensure_conversation(
-                session,
-                conv_id=conv_id,
-                org_id=org_id,
-                ws_id=ws_id,
-                user_id=user_id,
-            )
-            seen_conv.add(conv_id)
-        be = BillingEvent(
-            org_id=org_id,
-            workspace_id=ws_id,
-            user_id=user_id,
-            conversation_id=conv_id,
-            event_type="llm_call",
-            cost_amount_micro=int(r["cost_micro"]),
-            currency="USD",
-            started_at=r["started_at"],
-            ended_at=r["started_at"] + timedelta(milliseconds=200),
-            duration_ms=200,
-            status="ok",
-        )
-        le = LlmBillingEvent(
-            billing_event_id=be.id,
-            provider=str(r["provider"]),
-            model_id=str(r["model_id"]),
-            input_tokens=int(r.get("input", 0)),
-            output_tokens=int(r.get("output", 0)),
-            cache_read_tokens=int(r.get("cache_read", 0)),
-            cache_write_tokens=int(r.get("cache_write", 0)),
-        )
-        session.add(be)
-        session.add(le)
-    await session.commit()
 
 
 async def test_get_timeseries_workspace_two_workspaces_two_days() -> None:
@@ -320,174 +146,6 @@ async def test_get_timeseries_rank_by_tokens_with_zero_cost() -> None:
             p["input_tokens"] + p["output_tokens"] for s in series for p in s["points"]
         )
         assert total_tokens == sum(int(r["input"]) + int(r["output"]) for r in rows)
-
-
-async def test_cost_summary_returns_by_user(async_client: httpx.AsyncClient) -> None:
-    """/admin/cost/summary returns a `by_user` aggregation with the expected fields."""
-    async with _db_session() as session:
-        day = datetime(2026, 5, 5, 12, tzinfo=UTC)
-        await _seed_events(
-            session,
-            org_id=DEFAULT_ORG_ID,
-            rows=[
-                {
-                    "workspace_id": DEFAULT_WS_ID,
-                    "user_id": "usr-by-user-a",
-                    "user_display_name": "Alice Admin",
-                    "provider": "openai",
-                    "model_id": "gpt-4o",
-                    "started_at": day,
-                    "cost_micro": 1_500_000,
-                    "input": 120,
-                    "output": 30,
-                    "cache_read": 5,
-                    "cache_write": 7,
-                },
-                {
-                    "workspace_id": DEFAULT_WS_ID,
-                    "user_id": "usr-by-user-b",
-                    # No display_name → label falls back to email.
-                    "provider": "openai",
-                    "model_id": "gpt-4o",
-                    "started_at": day,
-                    "cost_micro": 750_000,
-                    "input": 60,
-                    "output": 12,
-                    "cache_read": 2,
-                    "cache_write": 3,
-                },
-            ],
-        )
-
-    resp = await async_client.get(
-        "/api/v1/admin/cost/summary",
-        params={"from_date": "2026-05-01", "to_date": "2026-05-31"},
-    )
-    assert resp.status_code == 200, resp.text
-    body = resp.json()
-
-    assert "by_user" in body, f"missing `by_user` key in response: {body.keys()}"
-    by_user = body["by_user"]
-    assert isinstance(by_user, list)
-    by_bucket = {row["bucket"]: row for row in by_user}
-    assert {"usr-by-user-a", "usr-by-user-b"}.issubset(by_bucket)
-
-    expected_fields = {
-        "bucket",
-        "bucket_type",
-        "bucket_label",
-        "input_tokens",
-        "output_tokens",
-        "cache_read_tokens",
-        "cache_write_tokens",
-        "cost_amount_micro",
-        "currency",
-        "call_count",
-    }
-    for row in by_user:
-        assert expected_fields.issubset(row.keys()), (
-            f"row missing fields: expected {expected_fields}, got {row.keys()}"
-        )
-        assert row["bucket_type"] == "user"
-
-    assert by_bucket["usr-by-user-a"]["bucket_label"] == "Alice Admin"
-    assert by_bucket["usr-by-user-b"]["bucket_label"] == "usr-by-user-b@test-billing-cost.local"
-
-    # Workspace label resolves from workspace.name (not the raw public id).
-    by_ws = {row["bucket"]: row for row in body["by_workspace"]}
-    assert DEFAULT_WS_ID in by_ws
-    ws_label = by_ws[DEFAULT_WS_ID]["bucket_label"]
-    assert ws_label is not None and ws_label != DEFAULT_WS_ID
-
-
-async def test_timeseries_workspace_happy_path(async_client: httpx.AsyncClient) -> None:
-    """Seed two workspaces, verify both appear unfiltered and only one with workspace_ids."""
-    ws_a = "ws-ts-hp-a"
-    ws_b = "ws-ts-hp-b"
-    async with _db_session() as session:
-        day = datetime(2026, 5, 5, 12, tzinfo=UTC)
-        await _seed_events(
-            session,
-            org_id=DEFAULT_ORG_ID,
-            rows=[
-                {
-                    "workspace_id": ws_a,
-                    "user_id": "usr-ts-hp-a",
-                    "provider": "openai",
-                    "model_id": "gpt-4o",
-                    "started_at": day,
-                    "cost_micro": 1_000_000,
-                    "input": 100,
-                    "output": 20,
-                },
-                {
-                    "workspace_id": ws_b,
-                    "user_id": "usr-ts-hp-b",
-                    "provider": "openai",
-                    "model_id": "gpt-4o",
-                    "started_at": day,
-                    "cost_micro": 500_000,
-                    "input": 50,
-                    "output": 10,
-                },
-            ],
-        )
-
-    # The default range is current-month-to-date; the seeded events live in
-    # 2026-05, so pin the query range explicitly to cover the seed date.
-    range_params = {"from_date": "2026-05-01", "to_date": "2026-05-31"}
-
-    # 1. Unfiltered: both workspace buckets present
-    resp = await async_client.get(
-        "/api/v1/admin/cost/timeseries",
-        params={"dimension": "workspace", "granularity": "day", **range_params},
-    )
-    assert resp.status_code == 200, resp.text
-    body = resp.json()
-    assert body["dimension"] == "workspace"
-    assert body["granularity"] == "day"
-    assert "series" in body and isinstance(body["series"], list)
-    by_bucket = {s["bucket"]: s for s in body["series"]}
-    assert {ws_a, ws_b}.issubset(by_bucket), f"expected both ws buckets, got {set(by_bucket)}"
-    assert by_bucket[ws_a]["bucket_label"] == f"Test {ws_a}"
-    assert by_bucket[ws_b]["bucket_label"] == f"Test {ws_b}"
-
-    # 2. Filtered by workspace_ids=ws_a: only ws_a bucket appears
-    resp = await async_client.get(
-        "/api/v1/admin/cost/timeseries",
-        params={
-            "dimension": "workspace",
-            "granularity": "day",
-            "workspace_ids": ws_a,
-            **range_params,
-        },
-    )
-    assert resp.status_code == 200, resp.text
-    body = resp.json()
-    filtered_buckets = {s["bucket"] for s in body["series"]}
-    assert filtered_buckets == {ws_a}, (
-        f"expected only {ws_a} bucket after filter, got {filtered_buckets}"
-    )
-
-
-async def test_timeseries_rejects_invalid_dimension(async_client: httpx.AsyncClient) -> None:
-    resp = await async_client.get(
-        "/api/v1/admin/cost/timeseries",
-        params={"dimension": "skill"},
-    )
-    assert resp.status_code == 422  # FastAPI validation
-
-
-async def test_timeseries_requires_admin(
-    member_client: tuple[httpx.AsyncClient, str],
-) -> None:
-    """Non-admin org member is rejected by the admin gate."""
-    client, _ws_id = member_client
-    resp = await client.get(
-        "/api/v1/admin/cost/timeseries",
-        params={"dimension": "workspace"},
-    )
-    assert resp.status_code == 403, resp.text
 
 
 async def test_get_timeseries_weekly_granularity_aggregates_days() -> None:
